refactor: 💡 split formatter test into categories
